requirements/dev.txt
-r prod.txt
black==23.9.1
django-debug-toolbar==4.2.0
parameterized==0.9.0